refactor : members와 security 의미상으로 기존 domain 단계에서 분
