+++ src/main/java/com/be/rebook/domain/members/controller/MembersController.java
@@ -2,10 +2,10 @@ package com.be.rebook.domain.members.controller;
 
 import com.be.rebook.domain.members.dto.UserinfoDTO;
 import com.be.rebook.domain.members.entity.Members;
-import com.be.rebook.domain.members.entity.RefreshTokens;
+import com.be.rebook.domain.security.entity.RefreshTokens;
 import com.be.rebook.domain.members.service.MemberService;
 import com.be.rebook.domain.members.dto.UpdateDTO;
-import com.be.rebook.domain.members.service.ReissueService;
+import com.be.rebook.domain.security.service.ReissueService;
 import com.be.rebook.global.config.BaseResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
