+++ src/main/java/com/be/rebook/domain/members/service/MemberService.java
@@ -4,15 +4,15 @@ import com.be.rebook.domain.members.dto.UpdateDTO;
 import com.be.rebook.domain.members.dto.UserinfoDTO;
 import com.be.rebook.domain.members.entity.Majors;
 import com.be.rebook.domain.members.entity.Members;
-import com.be.rebook.domain.members.entity.RefreshTokens;
+import com.be.rebook.domain.security.entity.RefreshTokens;
 import com.be.rebook.domain.members.entity.Universities;
 import com.be.rebook.domain.members.repository.MajorsRepository;
 import com.be.rebook.domain.members.repository.MembersRepository;
-import com.be.rebook.domain.members.repository.RefreshTokensRepository;
+import com.be.rebook.domain.security.repository.RefreshTokensRepository;
 import com.be.rebook.domain.members.repository.UniversitiesRepository;
 import com.be.rebook.global.exception.BaseException;
 import com.be.rebook.global.exception.ErrorCode;
-import com.be.rebook.global.security.jwt.JWTUtil;
+import com.be.rebook.domain.security.jwt.JWTUtil;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,7 +20,6 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class MemberService {
