+++ src/main/java/com/be/rebook/domain/security/dto/CustomUserDetails.java
@@ -1,4 +1,4 @@
-package com.be.rebook.domain.members.dto;
+package com.be.rebook.domain.security.dto;
 
 
 import com.be.rebook.domain.members.entity.Members;
