+++ src/main/java/com/be/rebook/domain/security/repository/RefreshTokensRepository.java
@@ -1,6 +1,6 @@
-package com.be.rebook.domain.members.repository;
+package com.be.rebook.domain.security.repository;
 
-import com.be.rebook.domain.members.entity.RefreshTokens;
+import com.be.rebook.domain.security.entity.RefreshTokens;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
