+++ src/main/java/com/be/rebook/domain/security/service/CustomUserDetailService.java
@@ -1,7 +1,6 @@
-package com.be.rebook.domain.members.service;
+package com.be.rebook.domain.security.service;
 
-import com.be.rebook.domain.members.dto.CustomUserDetails;
-import com.be.rebook.domain.members.entity.Members;
+import com.be.rebook.domain.security.dto.CustomUserDetails;
 import com.be.rebook.domain.members.repository.MembersRepository;
 import com.be.rebook.global.exception.BaseException;
 import com.be.rebook.global.exception.ErrorCode;
